fix: Account for whitespace when replacing operators and symbols
For example, replace ' and ' in "p and q" with ' \ \land \ ' rather than 'and' with '\ \land \'. This was intentional at first, but now we're preventing future problems if we want to implement features like replacing 'text(some text here)' with \text{some text here} etc. Really though I should have used regex from the start, but I didn't/don't have intentions for this to have a bunch of features. As usual though, I start with some small cli tool because I'm bored and then try to expand it beyond what it's base infrastructure should support.

This commit also removes replacing 't' and 'c' with '\mathbf{t}' etc. We
can implement *t* or something like that eventually.