feat: Copy to clipboard or overwrite/append to file in either input mode
If the user passes a statement, the output LaTeX will be the converted
statement. If the user passes an input file, the output LaTeX will be a
Markdown table with the converted LaTeX and laws.

If the user does not pass any additional args, the output LaTeX will be
copied to the system clipboard.
If the user passes the -o or -a flags, the output LaTeX will be
written/appended to the given output file. However, the -a flag will
only work with an existing output file. The user will still be asked to
confirm an overwrite if a file does exist when overwriting with -o.